+++ include/sequence_generator.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <wchar.h>
 
 enum sequence_gen_type
@@ -15,7 +16,7 @@ typedef struct sequence_generator
     unsigned int type;
     bool debug;
     void (*status)(void *self, wchar_t *wstring);
-    int (*generate)(void *self, void *data);
+    uint16_t (*generate)(void *self, void *data);
     void (*set_debug)(void *self, bool b);
     void (*event_notify)(void *self, unsigned int event_type);
 } sequence_generator;
